requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorize==1.1.0
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
dialogflow==0.6.0
dnspython==1.16.0
emoji==0.5.2
Emojipedia==0.4.0
entrypoints==0.3
feedparser==5.2.1
Flask==1.0.3
fuzzywuzzy==0.17.0
gnewsclient==1.12
google-api-core==1.11.1
google-auth==1.6.3
googleapis-common-protos==1.6.0
googletrans==2.4.0
grpcio==1.21.1
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
PTable==0.9.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.8.0
pyOpenSSL==19.0.0
pyrsistent==0.15.2
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.22.0
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
twilio==6.27.1
urllib3==1.25.3
wcwidth==0.1.7
weather-api==1.0.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wikipedia==1.4.0